refactor: Extract index loading logic from GallifreyDB::with_unified_config
Moved the massive index loading logic from `src/db.rs` into a new helper function `load_indexes_startup` in `src/storage/index_persistence/operations.rs`. This reduces the size of `src/db.rs` by ~270 lines and improves modularity.

- Extracted logic includes loading manifest, strings, graph index, temporal index, and vector indexes.
- Fixed some clippy warnings (collapsible ifs).
- Removed unused import in `src/db.rs`.
- Updated `load_vector_indexes` signature to be more idiomatic (takes reference instead of Arc).
